project/pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = ""
authors = [
    {name = "brian",email = "kimchanjoong54@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.8.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-allauth (>=65.11.2,<66.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "cryptography (>=46.0.1,<47.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "boto3 (>=1.40.33,<2.0.0)",
    "django-storages (>=1.14.6,<2.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "dj-rest-auth (>=7.0.1,<8.0.0)",
    "requests-oauthlib (>=2.0.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "project", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
